Dividing rewarbadle objects into separate files
